Javascript 파일 통합, Product Swiper 디자인 수정 [BugFix] : [디자인 오류 및 javascript 빌드 오류 수정] 1. Product Swiper 디자인이 안 맞는 부분을 수정했습니다. 2. 나눠져 있던 javascript 파일을 main.js에 통합했습니다.
-
