+++ index.html
@@ -10,9 +10,6 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <script type="module" src="/src/js/include.js"></script>
     <script type="module" src="/main.js"></script>
-    <script type="module" src="/main-product.js"></script>
-    <script type="module" src="/main-cProduct-swiper.js"></script>
-    <script type="module" src="/main-advertising.js"></script>
   </head>
 
   <body>
@@ -41,7 +38,7 @@
 
     <!-- 이 상품 어때요? 컨테이너 -->
     <div
-      class="product-container flex w-screen flex-nowrap justify-center py-40pxr"
+      class="product-container flex w-full flex-nowrap justify-center py-40pxr"
     >
       <div class="product-inner flex w-[1050px] flex-col">
         <h2 class="mb-28pxr inline-block text-center text-28pxr text-content">
@@ -60,10 +57,7 @@
     <!-- 팝업창 마크업 -->
     <dialog id="dialog" class="overflow-hidden rounded-[16px]">
       <a href="/">
-        <img
-          src="./src/assets/popup.png"
-          alt="도라에몽 비밀도구박물관 극장판"
-        />
+        <img src="/src/assets/popup.png" alt="도라에몽 비밀도구박물관 극장판" />
       </a>
       <form
         method="dialog"
@@ -91,7 +85,7 @@
 
     <!-- 도구를 다 보여주는 회전 목마 -->
     <div
-      class="product-container flex w-screen flex-nowrap justify-center py-40pxr"
+      class="product-container flex w-full flex-nowrap justify-center py-40pxr"
     >
       <div class="product-inner flex w-[1050px] flex-col">
         <h2 class="mb-28pxr inline-block text-center text-28pxr text-content">
