--- main-cProduct-swiper.js
@@ -1,34 +1,0 @@
-// import PocketBase from 'pocketbase';
-import Swiper from 'swiper/bundle';
-import './src/styles/tailwind.css';
-import 'swiper/css/bundle';
-
-// const pb = new PocketBase('import.meta.env.VITE_PB_URL');
-// const cProductImg = document.querySelector('.swiper-wrapper');
-
-// function getPbImageURL(collectionId, id, fileName = 'photo') {
-//   return `${import.meta.env.VITE_PB_URL}/files/${collectionId}/${id}/${fileName}`;
-// }
-
-// const swiper = new Swiper('.swiper', {
-//   // Optional parameters
-//   direction: 'vertical',
-//   // 보여지는 슬라이드 수
-//   slidesPerView: 2,
-//   // 센터 모드
-//   centeredSlides: true,
-//   // 해당 슬라이드 클릭시 슬라이드 위치로 이동
-//   slideToClickedSlide: true,
-//   // (false-스와이핑안됨)버튼으로만 슬라이드 조작이 가능
-//   allowTouchMove: true,
-//   // space bar로 이동
-//   pagination: {
-//     el: '.swiper-pagination',
-//     clickable: true,
-//   },
-//   // Navigation arrows
-//   navigation: {
-//     nextEl: '.swiper-button-next',
-//     prevEl: '.swiper-button-prev',
-//   },
-// });
