+++ public/components/header.html
@@ -77,7 +77,7 @@
           <!-- 로고 -->
           <div class="flex items-center gap-x-4">
             <h1 class="header-logo w-24">
-              <img src="/public/logo-emong.png" alt="마켓에몽 로고" />
+              <img src="/logo-emong.png" alt="마켓에몽 로고" />
             </h1>
             <a
               href="/index.html"
